tests/wpt/tests/html/rendering/non-replaced-elements/form-controls/placeholder-opacity-default.tentative.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Placeholder Test: opacity default value</title>
    <link rel="author" title="Karl Dubost" href="mailto:kdubost@mozilla.com"
    />
    <link rel="help" href="https://drafts.csswg.org/css-pseudo-4/#placeholder-pseudo"
    />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <input
      id="opacity"
      placeholder="::placeholder should have default opacity: 1"
    />
    <script>
      test(function () {
        var target = document.getElementById("opacity");
        assert_equals(getComputedStyle(target, '::placeholder').opacity, "1");
      }, "Default opacity value is '1'");
    </script>
  </body>
</html>
